Large changes, fill in a lot, fix some bugs

+++ arch/generic/src/unistd/syscall.c
@@ -1,26 +1,6 @@
-/*
- *  Copyright (C) 2019 James Larrowe
- *
- *  This file is part of Minilibc.
- *
- *  Minilibc is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  Minilibc is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with Minilibc.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#include <stdint.h>
 #include <errno.h>
 
-intptr_t __syscall(intptr_t syscall_num, ...)
+long __syscall(long syscall_num, ...)
 {
     (void)syscall_num;
     return -ENOSYS;
